src/app/components/nav-bar/nav-bar.component.ts
import { Component, HostListener } from '@angular/core';

@Component({
  selector: 'app-nav-bar',
  templateUrl: './nav-bar.component.html',
  styleUrls: ['./nav-bar.component.scss']
})
export class NavBarComponent {
  public isSmallScreen: boolean = false;
  public isMenuOpen: boolean = false;

  constructor() {
    this.checkScreenSize();
  }

  @HostListener('window:resize')
  checkScreenSize() {
    this.isSmallScreen = window.innerWidth < 950;
  }

  public toggleMenu() {
    this.isMenuOpen = !this.isMenuOpen;
  }
}